requirements.txt
gtts
openai-whisper
pyttsx3
scikit-learn
joblib
pandas
numpy
ollama
tqdm